Apply suggestions from code reviews
Delete PHASE_1.5_PLAN.md

Fix BookTitleLink alignment by adding missing icon styles

Add the base styled-icons default styles (display: inline-block,
vertical-align: middle, overflow: hidden) to the LeftArrow component
in BookBanner to maintain visual alignment parity with the previous
styled-icons implementation.

These styles were automatically applied by the styled-icons library
but need to be explicitly added when using inline SVG components.


Fix unused colorSchema parameter in ChevronLeftIconBase

Renamed colorSchema to _colorSchema to indicate it's intentionally unused.
This parameter is destructured to prevent it from being forwarded to the
SVG element (it's a styled-components styling prop), but ESLint would
flag it as unused. The underscore prefix follows the convention for
intentionally unused parameters.

Addresses Copilot's code review comment.


Copilot again

Disambiguate Times icons to resolve naming conflict

- Renamed components/icons/Times.tsx default export from `Times` to `TimesIcon`
  to avoid conflict with existing components/Times.tsx
- Updated imports in HighlightsHelpInfo.tsx and NudgeStudyTools/styles.ts
- Added clarifying comments to both icon files explaining when to use each:
  * components/Times.tsx: Custom 15x15px icon for UI close buttons (modals, toasts, sidebars)
  * components/icons/Times.tsx: Font Awesome Times icon for content components migrated from styled-icons

This resolves the ambiguous default export naming issue identified by Copilot.


Fix type safety issues in GoToTopButton and DotMenu

Address Copilot review comments from Review #17 and #18:

1. GoToTopButton.tsx:
   - Replace loose [key: string]: unknown with proper type safety
   - Extend Omit<React.ComponentPropsWithoutRef<typeof PlainButton>, ...>
   - Extract className from props and merge it into classNames()
   - This prevents className override issues and provides proper type checking

2. DotMenu.tsx:
   - Change toggle prop type from React.ReactNode to React.ReactElement
   - This matches the requirement of React.cloneElement used in Dropdown
   - Prevents runtime errors from passing non-element values


Snaps, remove styled-icons, FC -> function
