+++ package.json
@@ -39,7 +39,6 @@
     "reselect": "4.0.0",
     "scroll-to-element": "^2.0.3",
     "styled-components": "^4",
-    "styled-icons": "^9",
     "typesafe-actions": "^4.4.2",
     "typescript": "^4",
     "uuid": "~7.0",
