+++ src/app/components/Details.tsx
@@ -13,8 +13,8 @@ interface IconProps extends React.SVGAttributes<SVGSVGElement> {
 }
 
 /**
- * Expand icon (caret-right) for Details component - filled triangle pointing right.
- * SVG path from Font Awesome Free solid caret-right (https://fontawesome.com - MIT License)
+ * Expand icon (caret-right) for Details component.
+ * SVG path from Font Awesome Free (https://fontawesome.com - MIT License)
  *
  * Note: Wrapped with styled() to enable styled-components component selector references
  */
@@ -24,12 +24,11 @@ function ExpandIconBase({ className, ...props }: IconProps) {
       className={classNames('details-expand-icon', className)}
       viewBox="0 0 192 512"
       aria-hidden="true"
-      focusable="false"
       {...props}
     >
       <path
         fill="currentColor"
-        d="M0 384.662V127.338c0-17.818 21.543-26.741 34.142-14.142l128.662 128.662c7.81 7.81 7.81 20.474 0 28.284L34.142 398.804C21.543 411.404 0 402.48 0 384.662z"
+        d="M0 384.662V127.338c0-17.818 21.543-26.741 34.142-14.142l128.662 128.662c7.81 7.81 7.81 20.474 0 28.284L34.142 398.804C21.543 411.403 0 402.48 0 384.662z"
       />
     </svg>
   );
@@ -49,7 +48,6 @@ function CollapseIconBase({ className, ...props }: IconProps) {
       className={classNames('details-collapse-icon', className)}
       viewBox="0 0 320 512"
       aria-hidden="true"
-      focusable="false"
       {...props}
     >
       <path
