+++ src/app/components/GoToTopButton.tsx
@@ -8,32 +8,31 @@ import './GoToTopButton.css';
 
 /**
  * Icon component for the "scroll to top" button.
- * SVG path from Font Awesome Free angle-up icon (https://fontawesome.com - MIT License)
+ * SVG path from Font Awesome Free 5.15.4 (https://fontawesome.com - MIT License)
  */
 export function GoToTopIcon({ className, ...props }: React.SVGAttributes<SVGSVGElement>) {
   return (
     <svg
       className={classNames('go-to-top-icon', className)}
-      viewBox="0 0 384 512"
+      viewBox="0 0 320 512"
       aria-hidden="true"
-      focusable="false"
       {...props}
     >
       <path
         fill="currentColor"
-        d="M352 352c-8.188 0-16.38-3.125-22.62-9.375L192 205.3l-137.4 137.4c-12.5 12.5-32.75 12.5-45.25 0s-12.5-32.75 0-45.25l160-160c12.5-12.5 32.75-12.5 45.25 0l160 160c12.5 12.5 12.5 32.75 0 45.25C368.4 348.9 360.2 352 352 352z"
+        d="M177 159.7l136 136c9.4 9.4 9.4 24.6 0 33.9l-22.6 22.6c-9.4 9.4-24.6 9.4-33.9 0L160 255.9l-96.4 96.4c-9.4 9.4-24.6 9.4-33.9 0L7 329.7c-9.4-9.4-9.4-24.6 0-33.9l136-136c9.4-9.5 24.6-9.5 34-.1z"
       />
     </svg>
   );
 }
 
-type GoToTopButtonProps = Omit<React.ComponentProps<typeof PlainButton>, 'aria-label' | 'onClick'> & {
+interface GoToTopButtonProps extends Omit<React.ComponentPropsWithoutRef<typeof PlainButton>, 'aria-label' | 'onClick' | 'className'> {
   i18nAriaLabel: string;
   onClick: () => void;
-};
+  className?: string;
+}
 
-const GoToTopButton = (props: GoToTopButtonProps) => {
-  const { i18nAriaLabel, onClick, className, ...rest } = props;
+const GoToTopButton = ({ i18nAriaLabel, onClick, className, ...rest }: GoToTopButtonProps) => {
   const intl = useIntl();
 
   return (
