+++ src/app/content/components/NudgeStudyTools/styles.ts
@@ -1,38 +1,10 @@
-import React from 'react';
 import styled, { css } from 'styled-components/macro';
 import { PlainButton } from '../../../components/Button';
 import htmlMessage from '../../../components/htmlMessage';
+import { TimesIcon } from '../../../components/icons/Times';
 import theme from '../../../theme';
 import { remsToPx } from '../../../utils';
 import { arrowDesktopHeight, closeButtonMobileMargin, closeButtonSize, contentWidth } from './constants';
-
-interface IconProps extends React.SVGAttributes<SVGSVGElement> {
-  className?: string;
-}
-
-/**
- * Times icon component.
- * SVG path from Font Awesome Free (https://fontawesome.com - MIT License)
- *
- * Note: Wrapped with styled() to enable styled-components component selector references
- */
-function TimesBase({ className, ...props }: IconProps) {
-  return (
-    <svg
-      className={className}
-      viewBox="0 0 352 512"
-      aria-hidden="true"
-      {...props}
-    >
-      <path
-        fill="currentColor"
-        d="M242.72 256l100.07-100.07c12.28-12.28 12.28-32.19 0-44.48l-22.24-22.24c-12.28-12.28-32.19-12.28-44.48 0L176 189.28 75.93 89.21c-12.28-12.28-32.19-12.28-44.48 0L9.21 111.45c-12.28 12.28-12.28 32.19 0 44.48L109.28 256 9.21 356.07c-12.28 12.28-12.28 32.19 0 44.48l22.24 22.24c12.28 12.28 32.2 12.28 44.48 0L176 322.72l100.07 100.07c12.28 12.28 32.2 12.28 44.48 0l22.24-22.24c12.28-12.28 12.28-32.19 0-44.48L242.72 256z"
-      />
-    </svg>
-  );
-}
-
-const Times = styled(TimesBase)``;
 
 export const NudgeWrapper = styled.div`
   display: contents;
@@ -105,7 +77,7 @@ export const NudgeArrow = styled.img`
   `}
 `;
 
-export const NudgeCloseIcon = styled(Times)`
+export const NudgeCloseIcon = styled(TimesIcon)`
   width: 1.1rem;
   color: ${theme.color.text.white};
 `;
@@ -125,7 +97,6 @@ export const NudgeCloseButton = styled(PlainButton)`
   display: flex;
   align-items: center;
   justify-content: center;
-
   ${theme.breakpoints.mobile(css`
     left: auto;
     top: ${remsToPx(closeButtonMobileMargin)}px;
@@ -147,7 +118,6 @@ export const NudgeBackground = styled.div`
   height: 100%;
   z-index: ${theme.zIndex.nudgeOverlay};
   pointer-events: none;
-
   display: grid;
   ${(props: NudgeSpotlightPlacement) => `
     grid-template: ${props.top}px ${props.height}px 1fr / ${props.left}px ${props.width}px 1fr;
@@ -168,7 +138,6 @@ export const ClickBlocker = styled.div`
   pointer-events: all;
   background-color: ${theme.color.black};
   opacity: 0.9;
-
   grid-area: ${({area}: ClickBlockerProps) => area};
 `;
