+++ src/app/content/highlights/components/SummaryPopup/HighlightsHelpInfo.tsx
@@ -6,41 +6,14 @@ import { useAnalyticsEvent } from '../../../../../helpers/analytics';
 import { user as userSelector } from '../../../../auth/selectors';
 import { PlainButton } from '../../../../components/Button';
 import htmlMessage from '../../../../components/htmlMessage';
+import { TimesIcon } from '../../../../components/icons/Times';
 import { bodyCopyRegularStyle } from '../../../../components/Typography';
 import theme from '../../../../theme';
 import { assertWindow } from '../../../../utils';
 import { disablePrint } from '../../../components/utils/disablePrint';
 
-interface IconProps extends React.SVGAttributes<SVGSVGElement> {
-  className?: string;
-}
-
-/**
- * Times icon component.
- * SVG path from Font Awesome Free (https://fontawesome.com - MIT License)
- *
- * Note: Wrapped with styled() to enable styled-components component selector references
- */
-function TimesBase({ className, ...props }: IconProps) {
-  return (
-    <svg
-      className={className}
-      viewBox="0 0 352 512"
-      aria-hidden="true"
-      {...props}
-    >
-      <path
-        fill="currentColor"
-        d="M242.72 256l100.07-100.07c12.28-12.28 12.28-32.19 0-44.48l-22.24-22.24c-12.28-12.28-32.19-12.28-44.48 0L176 189.28 75.93 89.21c-12.28-12.28-32.19-12.28-44.48 0L9.21 111.45c-12.28 12.28-12.28 32.19 0 44.48L109.28 256 9.21 356.07c-12.28 12.28-12.28 32.19 0 44.48l22.24 22.24c12.28 12.28 32.2 12.28 44.48 0L176 322.72l100.07 100.07c12.28 12.28 32.2 12.28 44.48 0l22.24-22.24c12.28-12.28 12.28-32.19 0-44.48L242.72 256z"
-      />
-    </svg>
-  );
-}
-
-const Times = styled(TimesBase)``;
-
 // This is copied from CallToActionPopup > styles.tsx
-// Wher should we store this kind of functions?
+// Where should we store this kind of function?
 const slideInFromBottom = keyframes`
   0% {
     bottom: -100%;
@@ -78,7 +51,7 @@ const Wrapper = styled.div`
   ${disablePrint}
 `;
 
-export const CloseIcon = styled(Times)`
+export const CloseIcon = styled(TimesIcon)`
   color: ${theme.color.secondary.lightGray.darkest};
   width: 1.4rem;
 `;
